Added mysql-based functional tests
Run with `tox -e mysql`. It's configured to connect to
"mysql://root@localhost/quark_functional_tests".

RM11693
