+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27,flake8
+envlist = py27,flake8,mysql
 
 [testenv]
 setenv = VIRTUAL_ENV={envdir}
@@ -12,7 +12,7 @@ setenv = VIRTUAL_ENV={envdir}
          NO_EVENTLET=1
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
-commands = nosetests {posargs}
+commands = nosetests --exclude=mysql {posargs}
 
 [tox:jenkins]
 sitepackages = True
@@ -32,7 +32,10 @@ setenv = VIRTUAL_ENV={envdir}
          NOSE_COVER_MIN_PERCENTAGE=90
 deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
-commands = nosetests --cover-package=quark --cover-erase {posargs}
+commands = nosetests --exclude=mysql --cover-package=quark --cover-erase {posargs}
+
+[testenv:mysql]
+commands = nosetests --where=quark/tests/functional/mysql {posargs}
 
 [testenv:venv]
 commands = {posargs}
